Move Outcome to prototypes.
